+++ CHANGES.rst
@@ -13,11 +13,15 @@ unreleased
   :class:`~environment.Environment` enables it, in order to avoid a
   slow initial import. (`#765`_)
 - Python 2.6 and 3.3 are not supported anymore.
-- The `map` filter in async mode now automatically awaits
+- The ``map`` filter in async mode now automatically awaits
+- Added a new ``ChainableUndefined`` class to support getitem
+  and getattr on an undefined object. (`#977`_)
 - Add :class:`loaders.ImportLibResourceLoader` loader using
-  ``importlib.resource``, that does not require ``setuptools`` or ```` (`#970`_)
+  ``importlib.resource``, that does not require ``setuptools``
+  or ``pkg_resources`` (`#970`_)
 
 .. _#765: https://github.com/pallets/jinja/issues/765
+.. _#977: https://github.com/pallets/jinja/issues/977
 .. _#970: https://github.com/pallets/jinja/issues/970